refactor: drop unnecessary $state rune from getCookies
The $state rune only captured the initial value when returned, and
no caller relies on reactivity, so use a plain object. Rename the
file to cookies.ts since it no longer contains runes.
