+++ src/lib/cookies.ts
@@ -1,7 +1,7 @@
 import { browser } from '$app/environment';
 
 export function getCookies(): { [key: string]: string } {
-	let cookies = $state({});
+	let cookies: { [key: string]: string } = {};
 	if (browser) {
 		cookies = Object.fromEntries(document.cookie.split(';').map((s) => s.trim().split('=')));
 	}
